update ssrpro
add new feature

+++ newpackage/sil/luci-app-shadowsocksr-pro/root/etc/shadowsocksr/gen-gfwlist.sh
@@ -10,8 +10,12 @@ generate_china_banned()
 			refresh_cmd="wget -O /tmp/gfwlist.b64 http://iytc.net/tools/list.b64 2>/dev/null"
 		fi
 		eval $refresh_cmd
-		cat /tmp/gfwlist.b64 | base64 -d > /tmp/gfwlist.txt
-		rm -f /tmp/gfwlist.b64
+		if [ "$?" == "0" ]; then
+			cat /tmp/gfwlist.b64 | base64 -d > /tmp/gfwlist.txt
+			rm -f /tmp/gfwlist.b64
+		else
+
+		fi
 	fi
 
 	cat /tmp/gfwlist.txt | sort -u |
